requirements.txt
python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain==0.3.27
langchain-community==0.3.29
unstructured==0.18.14 # Document loading
markdown==3.9
onnxruntime==1.22.1
chromadb==0.5.0 # Vector storage
mistralai==1.9.10 # LLM
tiktoken==0.11.0  # For embeddings 
